+++ ADUtilsTests/RegisterableViewTest.swift
@@ -21,8 +21,10 @@ class TableViewDataSource : NSObject, UITableViewDataSource {
     }
 }
 
+class NibTableViewCell: UITableViewCell {}
 class TableViewHeader : UITableViewHeaderFooterView {}
 class TableViewFooter : UITableViewHeaderFooterView {}
+class NibTableHeaderView: UITableViewHeaderFooterView {}
 
 class CollectionViewDataSource : NSObject, UICollectionViewDataSource {
 
@@ -34,9 +36,25 @@ class CollectionViewDataSource : NSObject, UICollectionViewDataSource {
      */
 
     func registerCells(_ collectionView: UICollectionView) {
-        collectionView.register(cell: .class(UICollectionViewCell.self))
-        collectionView.register(header: .class(CollectionViewHeader.self))
-        collectionView.register(footer: .class(CollectionViewFooter.self))
+        collectionView.register(
+            cells: [
+                .class(UICollectionViewCell.self),
+                .class(CollectionViewCell.self),
+                .nib(NibCollectionViewCell.self)
+            ]
+        )
+        collectionView.register(
+            headers: [
+                .class(CollectionViewHeader.self),
+                .class(UICollectionReusableView.self)
+            ]
+        )
+        collectionView.register(
+            footers: [
+                .class(CollectionViewFooter.self),
+                .class(UICollectionReusableView.self)
+            ]
+        )
     }
 
     func numberOfSections(in collectionView: UICollectionView) -> Int {
@@ -66,6 +84,8 @@ class CollectionViewDataSource : NSObject, UICollectionViewDataSource {
     }
 }
 
+class CollectionViewCell: UICollectionViewCell {}
+class NibCollectionViewCell: UICollectionViewCell {}
 class CollectionViewHeader : UICollectionReusableView {}
 class CollectionViewFooter : UICollectionReusableView {}
 
@@ -85,20 +105,63 @@ class RegisterableViewTest: QuickSpec {
             }
 
             it("Should register cell class and get not nil instance back") {
-                let registerableView = RegisterableView.class(UITableViewCell.self)
-                tableView.register(cell: registerableView)
+                tableView.register(
+                    cells: [
+                        .class(UITableViewCell.self),
+                        .nib(NibTableViewCell.self)
+                    ]
+                )
 
                 let cell: UITableViewCell = tableView.dequeueCell(at: indexPath)
                 expect(cell).toNot(beNil())
+                let nibCell: NibTableViewCell = tableView.dequeueCell(at: indexPath)
+                expect(nibCell).toNot(beNil())
             }
 
             it("Should register header class and get not nil instance back") {
                 tableView.register(header: .class(TableViewHeader.self))
                 tableView.register(footer: .class(TableViewFooter.self))
+                tableView.register(header: .nib(NibTableHeaderView.self))
 
                 let header: TableViewHeader = tableView.dequeueHeader()
-                let footer: TableViewFooter = tableView.dequeueFooter()
                 expect(header).toNot(beNil())
+                let nibHeader: NibTableHeaderView = tableView.dequeueHeader()
+                expect(nibHeader).toNot(beNil())
+                let footer: TableViewFooter = tableView.dequeueFooter()
+                expect(footer).toNot(beNil())
+            }
+        }
+
+        describe("UICollectionView") {
+
+            var simpleDataSource: CollectionViewDataSource!
+            var collectionView: UICollectionView!
+            let indexPath = IndexPath(row: 0, section: 0)
+
+            beforeEach {
+                simpleDataSource = CollectionViewDataSource()
+                let layout = UICollectionViewFlowLayout()
+                let rect = CGRect(x: 0, y: 0, width: 320, height: 1000)
+                collectionView = UICollectionView(frame: rect, collectionViewLayout: layout)
+                collectionView.dataSource = simpleDataSource
+                simpleDataSource.registerCells(collectionView)
+                collectionView.reloadData()
+                collectionView.layoutIfNeeded()
+            }
+
+            it("should register cell class and get not nil instance back") {
+                let cell: UICollectionViewCell = collectionView.dequeueCell(at: indexPath)
+                expect(cell).toNot(beNil())
+
+                let nibCell: NibCollectionViewCell = collectionView.dequeueCell(at: indexPath)
+                expect(nibCell).toNot(beNil())
+            }
+
+            it("should register header class and get not instance back") {
+                let header: CollectionViewHeader = collectionView.dequeueHeader(at: indexPath)
+                expect(header).toNot(beNil())
+
+                let footer: CollectionViewFooter = collectionView.dequeueFooter(at: indexPath)
                 expect(footer).toNot(beNil())
             }
         }